# implement faster unit quaterion rotation